# Improve getopenbmccons performance (#5507)