--- a/xCAT-server/share/xcat/cons/openbmc
+++ b/xCAT-server/share/xcat/cons/openbmc
@@ -64,7 +64,7 @@ sub getans {
 my $cmdref = {
     command   => ["getopenbmccons"],
     arg       => ["text"],
-    noderange => [ $ARGV[0] ]
+    node => [ $ARGV[0] ]
 };
 acquire_lock();
 # avoid of congestion